Add landing page with navigation
